+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Support passing bindable parameters in an array
 - Improve the TypeScript types to allow auto inferring of the overloaded argument of `Promiser`
 - Add `SQLite.SQLValue` and `SQLite.SQLBindable` types for the data coming from or going to the DB
+- Expose `sqlite3Worker1Promiser` to user code
 
 # 1.0.0 2023-03-14